% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metasignatures.R
\name{bootstrapMaxMetaCorr}
\alias{bootstrapMaxMetaCorr}
\title{bootstrapMaxMetaCorr
For an input matrix, bootstraps the correlation of maximal disjoint metasignatures, i.e.
those of size floor(dim(mymat)\link{2}/2).}
\usage{
bootstrapMaxMetaCorr(mymat, iter = 100)
}
\arguments{
\item{mymat}{Numeric matrix of dimension features x samples}

\item{iter}{Numeric, the number of pairs of disjoint metasignatures to sample}
}
\value{
Numeric vector of length iter corresponding to the Pearson correlations of maximal disjoint
metasignatures sampled from the matrix.
}
\description{
bootstrapMaxMetaCorr
For an input matrix, bootstraps the correlation of maximal disjoint metasignatures, i.e.
those of size floor(dim(mymat)\link{2}/2).
}
