% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosineDimension.R
\name{cosineVsDim}
\alias{cosineVsDim}
\title{Calculate the mean and sd of cosine similarity of standard iid normal vectors of
specified length.}
\usage{
cosineVsDim(dims = seq(5, 100, 5), N = 1000)
}
\arguments{
\item{dims}{numeric vector, length values to evaluate. Default: seq(5, 100, 5)}

\item{N}{number of vectors to sample}
}
\value{
A dataframe with the dimension, mean, and standard deviation of the cosine
similarity of the sample, excluding the diagonal
}
\description{
Calculate the mean and sd of cosine similarity of standard iid normal vectors of
specified length.
}
