% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metasignatures.R
\name{ecdfPointwise}
\alias{ecdfPointwise}
\title{ecdfPointwise
Takes a numeric vector and computes an ecdf on user-input values of x. This is particularly
useful for overlaying multiple ECDFs at the same points.}
\usage{
ecdfPointwise(mydat, breaks)
}
\arguments{
\item{mydat}{numeric vector to compute an ECDF}

\item{breaks}{vector of points at which to compute the ECDF.}
}
\value{
A dataframe with columns x - the domain, the breaks; and y - the fraction of input
values less than x.
}
\description{
ecdfPointwise
Takes a numeric vector and computes an ecdf on user-input values of x. This is particularly
useful for overlaying multiple ECDFs at the same points.
}
