% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosineDimension.R
\name{getCosineDist}
\alias{getCosineDist}
\title{Get cosine similarity distribution - for IID normals}
\usage{
getCosineDist(mydim = 500, N = 1000, mkfig = 1)
}
\arguments{
\item{mydim}{numeric, dimension of the vector space}

\item{N}{numeric, number of elements to sample from the vector space}

\item{mkfig}{logical, whether to create a figure.}
}
\value{
the cosine similarity matrix of dimension N x N
}
\description{
Get cosine similarity distribution - for IID normals
}
