% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleNormalLength.R
\name{getGenNormalLength}
\alias{getGenNormalLength}
\title{Calculate the empirical Euclidean length of multivariate normal random vectors with eigenvalues sampled
from a gamma distribution and non-zero mean sampled from a normal distribution}
\usage{
getGenNormalLength(
  ndim = seq(100, 5000, 100),
  iter = 10,
  distiter = 10,
  sPar = c(1, 2, 1, 2),
  meanSd = 2
)
}
\arguments{
\item{ndim}{numeric, a vector of dimensions to sample. Default seq(100, 1000, 100)}

\item{iter}{Number of iterations to use for a particular dimension and eigenvalue distribution}

\item{distiter}{Number of eigenvalue distributions to run for each dimension}

\item{sPar}{Numeric, vector of gamma distribution hyperparameters for the eigenvalues. The eigenvalues
are sampled from a gamma distribution with shape parameter gamma(shape=sPar\link{1}, rate=sPar\link{2}). and rate
parameter gamma(shape=sPar\link{3}, rate=sPar\link{4})}

\item{meanSd}{Numeric, standard deviation of the normal distribution from which to sample means}
}
\value{
Dataframe with columns: ndim - dimension of vector, meanLen - average length of vectors, sdLen -
standard deviation of length, theoryLength - theoretical length predicted according to Central Limit Theorem, Smith et al.
}
\description{
Calculate the empirical Euclidean length of multivariate normal random vectors with eigenvalues sampled
from a gamma distribution and non-zero mean sampled from a normal distribution
}
