% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metasignatures.R
\name{getMetaCorr}
\alias{getMetaCorr}
\title{getMetaCorr
For an input matrix, compute the Pearson's correlation of metasignatures drawn from that matrix
as a function of the metasignature size k.}
\usage{
getMetaCorr(mymat, kmax, iter = 100, seqvals = c())
}
\arguments{
\item{mymat}{Numeric matrix of dimension features x samples}

\item{kmax}{Numeric, maximum value to use for metasignature size k.}

\item{iter}{Numeric, the number of pairs of disjoint metasignatures to sample for each value of
metasignature size k. Default = 100.}

\item{seqvals}{Numeric, optional specification of the metasignature size to use. Default is 1:kmax.}
}
\value{
A list of (1) cormat - a matrix with dimension kmax x iter, where each entry is the Pearson's
correlation of a metasignature of size k,
(2) cordf - a data frame with columns metasize, mean correlation, sd correlation summarizing the
correlation of metasignatures generated from the input matrix.
}
\description{
getMetaCorr
For an input matrix, compute the Pearson's correlation of metasignatures drawn from that matrix
as a function of the metasignature size k.
}
