% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metasignatures.R
\name{getMetasigs}
\alias{getMetasigs}
\title{Extract metasignatures
getMetasigs takes an input of a matrix of signatures, where the rows are features and the
columns are signatures.  It generates metasignatures by averaging k signatures together, where
k is the given parameter (default 1).}
\usage{
getMetasigs(mymat, k = 1, return2 = 1, returnk = 0)
}
\arguments{
\item{mymat}{Numeric matrix of dimension features x samples}

\item{k}{Numeric, number of signatures to sample and average without replacement}

\item{return2}{logical: if true, it returns a list of two disjoint metasignatures of size k.}

\item{returnk}{logical: if true and return2 is false, returns as many disjoint
metasignatures of size k as possible. If return2 and returnk are both false, returns one metasignature}
}
\value{
some number of metasignatures depending on parameters specified
}
\description{
Extract metasignatures
getMetasigs takes an input of a matrix of signatures, where the rows are features and the
columns are signatures.  It generates metasignatures by averaging k signatures together, where
k is the given parameter (default 1).
}
