% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metasignatures.R
\name{getMetaSim}
\alias{getMetaSim}
\title{getMetaSim
For an input matrix, compute the similarity of metasignatures drawn from that matrix
as a function of the metasignature size k. Analogous to getMetaCorr, but allows Pearson,
Spearman, cosine similarity, and L1000's weighted connectivity score (wtcs).}
\usage{
getMetaSim(mymat, kmax, iter = 100, metric = "pearson", seqvals = c())
}
\arguments{
\item{mymat}{Numeric matrix of dimension features x samples}

\item{kmax}{Numeric, maximum value to use for metasignature size k.}

\item{iter}{Numeric, number of pairs of disjoint metasignatures to sample for each value of
metasignature size k. Default = 100.}

\item{metric}{One of "pearson", "spearman", "wtcs", or "cosine".}

\item{seqvals}{Numeric, optional specification of the metasignature size to use. Default is 1:kmax.}
}
\value{
List of: simmat - similarity matrix of dimension kmax x iter with the similarities
of the metasignatures; cordf - data frame summarizing the mean and standard deviation of the
similarities; metric - string indicating which metric was used.
}
\description{
getMetaSim
For an input matrix, compute the similarity of metasignatures drawn from that matrix
as a function of the metasignature size k. Analogous to getMetaCorr, but allows Pearson,
Spearman, cosine similarity, and L1000's weighted connectivity score (wtcs).
}
