% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metasignatures.R
\name{getMetaSimDs}
\alias{getMetaSimDs}
\title{getMetaSimDs
Apply getMetaSim to each group within a dataset.}
\usage{
getMetaSimDs(
  mymat,
  groupings,
  kmax = 1000,
  iter = 100,
  metric = "pearson",
  seqvals = c()
)
}
\arguments{
\item{mymat}{Numeric matrix of dimension features x samples}

\item{groupings}{Vector of class labels, either character or numeric. For each class, metacorrelation
is computed.}

\item{kmax}{Numeric, maximum value to use for metasignature size k (default 1000). For any group with
fewer than 2*kmax signatures, computes up to floor(groupsize/2).}

\item{iter}{Numeric, number of pairs of disjoint metasignatures to sample for each value of metasize k.
Default  = 100.}

\item{metric}{One of "pearson" (default), "spearman", "wtcs", or "cosine".}

\item{seqvals}{Numeric, optional specification of the metasignature size to use. Default is 1:kmax.}
}
\description{
getMetaSimDs
Apply getMetaSim to each group within a dataset.
}
