% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMoments.R
\name{getMoments}
\alias{getMoments}
\title{Calculate moments of a dataset}
\usage{
getMoments(mymat, outpath = "./moments.rds", rownames = c())
}
\arguments{
\item{mymat}{Numeric matrix of dimension features x samples}

\item{outpath}{Output path for the rds object with the calculated data. Default: ./moments.rds}

\item{rownames}{Optional character vector of character identifiers for the rows of the matrix.}
}
\value{
A dataframe with row identifiers, the moments, and the normality test
}
\description{
getMoments takes an input of a matrix of signatures, where the rows are features and the
columns are signatures.  It then calculates the mean, variance, and skewness of each row and applies
a test for normality.
}
