% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1kMetaCalc.R
\name{l1kMetaPosFrac}
\alias{l1kMetaPosFrac}
\title{l1kMetaSigs}
\usage{
l1kMetaPosFrac(dspath, metapath, outpath = ".")
}
\arguments{
\item{dspath}{Path to level5 control dataset}

\item{metapath}{Path to directory containing L1000 metadata}

\item{outpath}{Path for output files}
}
\description{
Extract metasignatures of various values to illustrate the tendencies of genes.
}
