% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1kMetaCalc.R
\name{l1kNullCalc}
\alias{l1kNullCalc}
\title{l1kNullCalc
Computes correlation of metasignatures of negative control DMSOs}
\usage{
l1kNullCalc(
  dspath = ".",
  metapath = ".",
  outpath = ".",
  kmax = 100,
  iter = 100,
  metric = "pearson",
  renorm = 0
)
}
\arguments{
\item{dspath}{Path to level5 control dataset}

\item{metapath}{Path to directory containing L1000 metadata}

\item{outpath}{Path for output files}

\item{kmax}{Maximum metasignature size to use}

\item{iter}{Number of iterations for each value of k}

\item{metric}{Metric to use in getMetaSims, default is "pearson"}

\item{renorm}{Whether to renormalize the data, currently not implemented. Default 0.}
}
\value{
Saves output, returns list of getMetaSimDs objects.
}
\description{
l1kNullCalc
Computes correlation of metasignatures of negative control DMSOs
}
