% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1kMetaCalc.R
\name{l1kSimDists}
\alias{l1kSimDists}
\title{l1kSimDists}
\usage{
l1kSimDists(dspath = ".", metapath = ".", outpath = ".")
}
\arguments{
\item{dspath}{Path to level5 control dataset}

\item{metapath}{Path to directory containing L1000 metadata}

\item{outpath}{Path for output files}
}
\value{
list of null getMetaSimDS and compound-specific getMetaSimDS outputs.
}
\description{
Compute similarity distributions and fraction of gene values that are positive for L1000
data to illustrate the metasignature anomaly.
}
