% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renormalize.R
\name{renormData}
\alias{renormData}
\title{renormData
This function takes a matrix and a subset of the columns, and renormalizes them using
one of several methods. The default is to center the matrix, so the mean is 0. It then
applies the transformation to a second matrix and returns the renormalized second matrix.
In practice, the second matrix could be the same as the first matrix, in which case
it returns the renormalized matrix.}
\usage{
renormData(mat1, mat2, colix = c(), method = "center")
}
\arguments{
\item{mat1}{The matrix to center, of dimension features x samples}

\item{mat2}{The matrix to return, of dimension features x samples}

\item{colix}{Numeric, the indices of columns to use (default is all).}

\item{method}{Method to use (default = center).}
}
\value{
A matrix, a renormalized mat2 per the criterion calculated on mat1.
}
\description{
renormData
This function takes a matrix and a subset of the columns, and renormalizes them using
one of several methods. The default is to center the matrix, so the mean is 0. It then
applies the transformation to a second matrix and returns the renormalized second matrix.
In practice, the second matrix could be the same as the first matrix, in which case
it returns the renormalized matrix.
}
