% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosineVariance.R
\name{sampleVarCos}
\alias{sampleVarCos}
\title{Sample the variance of cosine similarity for vectors drawn from a multivariate normal
distribution with eigenvalues drawn from a Gamma distribution.}
\usage{
sampleVarCos(
  ndim = seq(100, 1000, 100),
  iter = 3,
  distiter = 10,
  sPar = c(1, 2, 1, 2)
)
}
\arguments{
\item{ndim}{numeric, a vector of dimensions to sample. Default seq(100, 1000, 100)}

\item{iter}{Number of iterations to use for a particular dimension and eigenvalue distribution}

\item{distiter}{Number of eigenvalue distributions to run for each dimension}

\item{sPar}{Numeric, vector of gamma distribution hyperparameters for the eigenvalues. The eigenvalues
are sampled from a gamma distribution with shape parameter gamma(shape=sPar\link{1}, rate=sPar\link{2}). and rate
parameter gamma(shape=sPar\link{3}, rate=sPar\link{4})}
}
\value{
Dataframe with columns: dimension, theoretical variance based on eigenvalues, theoretical variance
on observed eigenvalues, and empirical variance.
}
\description{
Sample the variance of cosine similarity for vectors drawn from a multivariate normal
distribution with eigenvalues drawn from a Gamma distribution.
}
