% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1kMetaCalc.R
\name{l1kBgCalc}
\alias{l1kBgCalc}
\title{l1kBgCalc
Compute correlation of background metasignatures for L1000 cell lines}
\usage{
l1kBgCalc(
  datapath = ".",
  metapath = ".",
  outpath = ".",
  kmax = 100,
  iter = 100,
  renorm = 0,
  metric = "pearson"
)
}
\arguments{
\item{datapath}{Path to directory containing level5 ds}

\item{metapath}{Path to directory containing metadata}

\item{outpath}{Path for output files}

\item{kmax}{Maximum metasignature size to use}

\item{iter}{Number of iterations for each value of k}

\item{renorm}{Whether to renormalize the data, currently not implemented. Default 0.}

\item{metric}{Metric to use in getMetaSims. Default is Pearson.}
}
\value{
Saves output, doesn't return anything.
}
\description{
l1kBgCalc
Compute correlation of background metasignatures for L1000 cell lines
}
