% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1kMetaCalc.R
\name{l1kMetaCalc}
\alias{l1kMetaCalc}
\title{l1kMetaCalc
Compute correlation of metasignatures for L1000 cell lines}
\usage{
l1kMetaCalc(
  datapath = ".",
  metapath = ".",
  outpath = ".",
  kmax = 100,
  iter = 100
)
}
\arguments{
\item{datapath}{Path to directory containing level5 ds}

\item{metapath}{Path to directory containing metadata}

\item{outpath}{Path for output files}

\item{kmax}{Maximum metasignature size to use}

\item{iter}{Number of iterations for each value of k}
}
\value{
Saves output, doesn't return anything.
}
\description{
l1kMetaCalc
Compute correlation of metasignatures for L1000 cell lines
}
