% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosineVariance.R
\name{sampleVarCos2D}
\alias{sampleVarCos2D}
\title{Calculate the variance of cosine specifically for the two-dimensional case}
\usage{
sampleVarCos2D(myn = 1000, iter = 100, meanvec = c(0, 0))
}
\arguments{
\item{myn}{\itemize{
\item Number of vectors to sample at each ratio
}}

\item{iter}{\itemize{
\item Number of different values of u to sample on \link{0, 0.5}
}}

\item{meanvec}{\itemize{
\item Mean values to use (default (0,0)).
}}
}
\value{
dataframe with mean, variance, and approximation variance
}
\description{
Calculate the variance of cosine specifically for the two-dimensional case
}
